[NFI]: Split KrnlToAffine.cpp
