+++ src/Conversion/KrnlToAffine/CMakeLists.txt
@@ -1,7 +1,15 @@
 # SPDX-License-Identifier: Apache-2.0
 
 add_onnx_mlir_library(OMKrnlToAffine
-  KrnlToAffine.cpp
+  ConvertKrnlToAffine.cpp
+  KrnlCopyFromBuffer.cpp
+  KrnlCopyToBuffer.cpp
+  KrnlLoad.cpp
+  KrnlMatmul.cpp
+  KrnlMemset.cpp
+  KrnlStore.cpp
+  KrnlTerminator.cpp
+  KrnlToAffineHelper.cpp
 
   LINK_LIBS PUBLIC
   OMSpecializedKernelOpInterface
