+++ src/Dialect/Krnl/KrnlOps.cpp
@@ -35,6 +35,7 @@
 #include "src/Dialect/Krnl/KrnlOps.hpp"
 
 using namespace mlir;
+using namespace onnx_mlir;
 
 //===----------------------------------------------------------------------===//
 // KrnlOpsDialect
@@ -42,7 +43,7 @@ using namespace mlir;
 
 KrnlOpsDialect::KrnlOpsDialect(MLIRContext *context)
     : Dialect(getDialectNamespace(), context, TypeID::get<KrnlOpsDialect>()) {
-  addTypes<LoopType, StringType>();
+  addTypes<krnl::LoopType, krnl::StringType>();
   addOperations<
 #define GET_OP_LIST
 #include "src/Dialect/Krnl/KrnlOps.cpp.inc"
@@ -56,9 +57,9 @@ Type KrnlOpsDialect::parseType(DialectAsmParser &parser) const {
   MLIRContext *context = getContext();
 
   if (keyword == "loop")
-    return LoopType::get(context);
+    return krnl::LoopType::get(context);
   if (keyword == "string")
-    return StringType::get(context);
+    return krnl::StringType::get(context);
 
   parser.emitError(parser.getNameLoc(), "unknown krnl type: " + keyword);
   return Type();
@@ -66,8 +67,8 @@ Type KrnlOpsDialect::parseType(DialectAsmParser &parser) const {
 
 void KrnlOpsDialect::printType(Type type, DialectAsmPrinter &os) const {
   TypeSwitch<Type>(type)
-      .Case<LoopType>([&](Type) { os << "loop"; })
-      .Case<StringType>([&](Type) { os << "string"; })
+      .Case<krnl::LoopType>([&](Type) { os << "loop"; })
+      .Case<krnl::StringType>([&](Type) { os << "string"; })
       .Default([](Type) { llvm_unreachable("Unexpected 'krnl' type kind"); });
 }
 
@@ -81,7 +82,7 @@ void KrnlDefineLoopsOp::build(
     OpBuilder &builder, OperationState &result, int64_t num_loops) {
   // Create the same number of dimension handlers as the number of
   // dimensions in the associated integer set.
-  result.types.append(num_loops, LoopType::get(builder.getContext()));
+  result.types.append(num_loops, krnl::LoopType::get(builder.getContext()));
   result.addAttribute(
       getNumLoopsAttrName(), builder.getI64IntegerAttr(num_loops));
 }
@@ -102,8 +103,9 @@ ParseResult KrnlDefineLoopsOp::parse(
           KrnlDefineLoopsOp::getNumLoopsAttrName(), result.attributes))
     return failure();
 
-  auto loopTypes = llvm::SmallVector<Type, 4>(
-      numLoops.getValue().getSExtValue(), LoopType::get(builder.getContext()));
+  auto loopTypes =
+      llvm::SmallVector<Type, 4>(numLoops.getValue().getSExtValue(),
+          krnl::LoopType::get(builder.getContext()));
   if (parser.addTypesToList(loopTypes, result.types))
     return failure();
   return success();
@@ -259,7 +261,7 @@ ParseResult KrnlIterateOp::parse(OpAsmParser &parser, OperationState &result) {
   if (parser.parseOperandList(
           optimizedLoopRefs, OpAsmParser::Delimiter::Paren) ||
       parser.resolveOperands(optimizedLoopRefs,
-          LoopType::get(result.getContext()), result.operands))
+          krnl::LoopType::get(result.getContext()), result.operands))
     return failure();
 
   // Record how many optimized loops did we parse.
@@ -341,7 +343,7 @@ ParseResult KrnlIterateOp::parse(OpAsmParser &parser, OperationState &result) {
 
   while (failed(parser.parseOptionalRParen())) {
     // Parse an input loop operand;
-    operandParser.ParseOperand(LoopType::get(context), result.operands);
+    operandParser.ParseOperand(krnl::LoopType::get(context), result.operands);
     parser.parseArrow();
 
     // Parse induction variable.
@@ -431,7 +433,8 @@ void KrnlInstrumentOp::build(mlir::OpBuilder &builder, OperationState &state,
 
 void KrnlBlockOp::build(::mlir::OpBuilder &odsBuilder,
     ::mlir::OperationState &odsState, Value odsLoop, int64_t odsTileSize) {
-  SmallVector<Type, 4> blockResType(2, LoopType::get(odsBuilder.getContext()));
+  SmallVector<Type, 4> blockResType(
+      2, krnl::LoopType::get(odsBuilder.getContext()));
   build(odsBuilder, odsState, blockResType, odsLoop,
       odsBuilder.getI64IntegerAttr(odsTileSize));
 }
